remove unused liveviews + fix failing tests after changing instructor dashboard routes
