+++ test/oli_web/live/delivery/instructor_dashboard/assignments_tab_test.exs
@@ -1,4 +1,4 @@
-defmodule OliWeb.Delivery.InstructorDashboard.ManageLiveTest do
+defmodule OliWeb.Delivery.InstructorDashboard.AssignmentsTabTest do
   use ExUnit.Case, async: true
   use OliWeb.ConnCase
 
@@ -8,11 +8,12 @@ defmodule OliWeb.Delivery.InstructorDashboard.ManageLiveTest do
   alias Lti_1p3.Tool.ContextRoles
   alias Oli.Delivery.Sections
 
-  defp live_view_manage_route(section_slug) do
+  defp live_view_assignments_route(section_slug) do
     Routes.live_path(
       OliWeb.Endpoint,
-      OliWeb.Delivery.InstructorDashboard.ManageLive,
-      section_slug
+      OliWeb.Delivery.InstructorDashboard.InstructorDashboardLive,
+      section_slug,
+      :assignments
     )
   end
 
@@ -20,10 +21,11 @@ defmodule OliWeb.Delivery.InstructorDashboard.ManageLiveTest do
     test "can not access page when it is not logged in", %{conn: conn} do
       section = insert(:section)
 
-      redirect_path = "/session/new?request_path=%2Fsections%2F#{section.slug}%2Fmanage"
+      redirect_path =
+        "/session/new?request_path=%2Fsections%2F#{section.slug}%2Finstructor_dashboard%2Fassignments"
 
       assert {:error, {:redirect, %{to: ^redirect_path}}} =
-               live(conn, live_view_manage_route(section.slug))
+               live(conn, live_view_assignments_route(section.slug))
     end
   end
 
@@ -37,7 +39,7 @@ defmodule OliWeb.Delivery.InstructorDashboard.ManageLiveTest do
       redirect_path = "/unauthorized"
 
       assert {:error, {:redirect, %{to: ^redirect_path}}} =
-               live(conn, live_view_manage_route(section.slug))
+               live(conn, live_view_assignments_route(section.slug))
     end
   end
 
@@ -48,7 +50,7 @@ defmodule OliWeb.Delivery.InstructorDashboard.ManageLiveTest do
       redirect_path = "/unauthorized"
 
       assert {:error, {:redirect, %{to: ^redirect_path}}} =
-               live(conn, live_view_manage_route(section.slug))
+               live(conn, live_view_assignments_route(section.slug))
     end
 
     test "can access page if enrolled to section", %{
@@ -58,20 +60,17 @@ defmodule OliWeb.Delivery.InstructorDashboard.ManageLiveTest do
     } do
       Sections.enroll(instructor.id, section.id, [ContextRoles.get_role(:context_instructor)])
 
-      {:ok, view, _html} = live(conn, live_view_manage_route(section.slug))
+      {:ok, view, _html} = live(conn, live_view_assignments_route(section.slug))
 
-      # Manage tab is the selected one
+      # Assignments tab is the selected one
       assert has_element?(
                view,
-               ~s{a[href="#{live_view_manage_route(section.slug)}"].border-b-2},
-               "Manage"
+               ~s{a[href="#{live_view_assignments_route(section.slug)}"].border-b-2},
+               "Assignments"
              )
 
-      # Manage tab content gets rendered
-      assert has_element?(view, ~s{div[id="overview"]})
-
-      # Collab Space Group gets rendered
-      assert render(view) =~ "Collaboration Space"
+      # Assignments tab content gets rendered
+      assert has_element?(view, ~s{p}, "Not available yet")
     end
   end
 end
