+++ test/oli_web/live/delivery/instructor_dashboard/content_tab_test.exs
@@ -1,4 +1,4 @@
-defmodule OliWeb.Delivery.InstructorDashboard.AssignmentsLiveTest do
+defmodule OliWeb.Delivery.InstructorDashboard.ContentTabTest do
   use ExUnit.Case, async: true
   use OliWeb.ConnCase
 
@@ -8,11 +8,12 @@ defmodule OliWeb.Delivery.InstructorDashboard.AssignmentsLiveTest do
   alias Lti_1p3.Tool.ContextRoles
   alias Oli.Delivery.Sections
 
-  defp live_view_assignments_route(section_slug) do
+  defp live_view_content_route(section_slug) do
     Routes.live_path(
       OliWeb.Endpoint,
-      OliWeb.Delivery.InstructorDashboard.AssignmentsLive,
-      section_slug
+      OliWeb.Delivery.InstructorDashboard.InstructorDashboardLive,
+      section_slug,
+      :content
     )
   end
 
@@ -20,10 +21,11 @@ defmodule OliWeb.Delivery.InstructorDashboard.AssignmentsLiveTest do
     test "can not access page when it is not logged in", %{conn: conn} do
       section = insert(:section)
 
-      redirect_path = "/session/new?request_path=%2Fsections%2F#{section.slug}%2Fassignments"
+      redirect_path =
+        "/session/new?request_path=%2Fsections%2F#{section.slug}%2Finstructor_dashboard%2Fcontent"
 
       assert {:error, {:redirect, %{to: ^redirect_path}}} =
-               live(conn, live_view_assignments_route(section.slug))
+               live(conn, live_view_content_route(section.slug))
     end
   end
 
@@ -37,7 +39,7 @@ defmodule OliWeb.Delivery.InstructorDashboard.AssignmentsLiveTest do
       redirect_path = "/unauthorized"
 
       assert {:error, {:redirect, %{to: ^redirect_path}}} =
-               live(conn, live_view_assignments_route(section.slug))
+               live(conn, live_view_content_route(section.slug))
     end
   end
 
@@ -48,7 +50,7 @@ defmodule OliWeb.Delivery.InstructorDashboard.AssignmentsLiveTest do
       redirect_path = "/unauthorized"
 
       assert {:error, {:redirect, %{to: ^redirect_path}}} =
-               live(conn, live_view_assignments_route(section.slug))
+               live(conn, live_view_content_route(section.slug))
     end
 
     test "can access page if enrolled to section", %{
@@ -58,17 +60,17 @@ defmodule OliWeb.Delivery.InstructorDashboard.AssignmentsLiveTest do
     } do
       Sections.enroll(instructor.id, section.id, [ContextRoles.get_role(:context_instructor)])
 
-      {:ok, view, _html} = live(conn, live_view_assignments_route(section.slug))
+      {:ok, view, _html} = live(conn, live_view_content_route(section.slug))
 
-      # Assignments tab is the selected one
+      # Module tab is the selected one
       assert has_element?(
                view,
-               ~s{a[href="#{live_view_assignments_route(section.slug)}"].border-b-2},
-               "Assignments"
+               ~s{a[href="#{live_view_content_route(section.slug)}"].border-b-2},
+               "Modules"
              )
 
-      # Assignments tab content gets rendered
-      assert has_element?(view, ~s{p}, "Not available yet")
+      # Module tab content gets rendered
+      assert has_element?(view, ~s{h5}, "Course Overview")
     end
   end
 end
