+++ test/oli_web/live/delivery/instructor_dashboard/learning_objectives_tab_test.exs
@@ -1,4 +1,4 @@
-defmodule OliWeb.Delivery.InstructorDashboard.ContentLiveTest do
+defmodule OliWeb.Delivery.InstructorDashboard.LearningObjectivesTabTest do
   use ExUnit.Case, async: true
   use OliWeb.ConnCase
 
@@ -8,11 +8,12 @@ defmodule OliWeb.Delivery.InstructorDashboard.ContentLiveTest do
   alias Lti_1p3.Tool.ContextRoles
   alias Oli.Delivery.Sections
 
-  defp live_view_content_route(section_slug) do
+  defp live_view_learning_objectives_route(section_slug) do
     Routes.live_path(
       OliWeb.Endpoint,
-      OliWeb.Delivery.InstructorDashboard.ContentLive,
-      section_slug
+      OliWeb.Delivery.InstructorDashboard.InstructorDashboardLive,
+      section_slug,
+      :learning_objectives
     )
   end
 
@@ -20,10 +21,11 @@ defmodule OliWeb.Delivery.InstructorDashboard.ContentLiveTest do
     test "can not access page when it is not logged in", %{conn: conn} do
       section = insert(:section)
 
-      redirect_path = "/session/new?request_path=%2Fsections%2F#{section.slug}%2Fcontent"
+      redirect_path =
+        "/session/new?request_path=%2Fsections%2F#{section.slug}%2Finstructor_dashboard%2Flearning_objectives"
 
       assert {:error, {:redirect, %{to: ^redirect_path}}} =
-               live(conn, live_view_content_route(section.slug))
+               live(conn, live_view_learning_objectives_route(section.slug))
     end
   end
 
@@ -37,7 +39,7 @@ defmodule OliWeb.Delivery.InstructorDashboard.ContentLiveTest do
       redirect_path = "/unauthorized"
 
       assert {:error, {:redirect, %{to: ^redirect_path}}} =
-               live(conn, live_view_content_route(section.slug))
+               live(conn, live_view_learning_objectives_route(section.slug))
     end
   end
 
@@ -48,7 +50,7 @@ defmodule OliWeb.Delivery.InstructorDashboard.ContentLiveTest do
       redirect_path = "/unauthorized"
 
       assert {:error, {:redirect, %{to: ^redirect_path}}} =
-               live(conn, live_view_content_route(section.slug))
+               live(conn, live_view_learning_objectives_route(section.slug))
     end
 
     test "can access page if enrolled to section", %{
@@ -57,18 +59,17 @@ defmodule OliWeb.Delivery.InstructorDashboard.ContentLiveTest do
       conn: conn
     } do
       Sections.enroll(instructor.id, section.id, [ContextRoles.get_role(:context_instructor)])
+      {:ok, view, _html} = live(conn, live_view_learning_objectives_route(section.slug))
 
-      {:ok, view, _html} = live(conn, live_view_content_route(section.slug))
-
-      # Module tab is the selected one
+      # LearningObjectives tab is the selected one
       assert has_element?(
                view,
-               ~s{a[href="#{live_view_content_route(section.slug)}"].border-b-2},
-               "Modules"
+               ~s{a[href="#{live_view_learning_objectives_route(section.slug)}"].border-b-2},
+               "Learning Objectives"
              )
 
-      # Module tab content gets rendered
-      assert has_element?(view, ~s{h5}, "Course Overview")
+      # LearningObjectives tab content gets rendered
+      assert has_element?(view, ~s{p}, "Not available yet")
     end
   end
 end
